+++ src/main/java/site/dogether/member/service/MemberService.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import site.dogether.auth.controller.request.LoginRequest;
 import site.dogether.auth.controller.request.WithdrawRequest;
+import site.dogether.auth.infrastructure.AppleOAuthProvider;
 import site.dogether.auth.infrastructure.JwtHandler;
-import site.dogether.auth.infrastructure.client.apple.AppleOAuthProvider;
 import site.dogether.member.domain.Member;
 import site.dogether.member.infrastructure.entity.MemberJpaEntity;
 import site.dogether.member.infrastructure.repository.MemberJpaRepository;
